Move FilteredPersonListChangedEvent from UI to Model
It was wrongly placed in the UI folder instead of the Model Folder
